Moved more "Properties" components -> "Properties"
